+++ src/index.ts
@@ -1,4 +1,4 @@
-export { IonicButtonComponent } from './ionic-button.component';
+export { IonicButtonComponent } from './widgets/button/ionic-button.component';
 export { IonicInputComponent } from './widgets/input/ionic-input.component';
 export { IonicCheckboxComponent } from './widgets/checkbox/ionic-checkbox.component';
 export { IonicFrameworkComponent } from './ionic-framework.component';
