+++ src/ionic-framework.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { JsonSchemaFormService, WidgetLibraryModule } from 'angular2-json-schema-form';
 import { IonicModule } from 'ionic-angular';
-import { IonicButtonComponent } from './ionic-button.component';
+import { IonicButtonComponent } from './widgets/button/ionic-button.component';
 import { IonicInputComponent } from './widgets/input/ionic-input.component';
 import { IonicCheckboxComponent } from './widgets/checkbox/ionic-checkbox.component'
 import { IonicFrameworkComponent } from './ionic-framework.component';
