+++ src/widgets/button/ionic-button.template.html
@@ -1,0 +1,8 @@
+<div [class]="options?.htmlClass || ''">
+    <button ion-button [attr.readonly]="options?.readonly ? 'readonly' : null" [attr.aria-describedby]="'control' + layoutNode?._id + 'Status'"
+        [color]="options?.color || 'primary'" [disabled]="controlDisabled || options?.readonly" [id]="'control' + layoutNode?._id"
+        [name]="controlName" [type]="layoutNode?.type" [value]="controlValue" (click)="updateValue($event)">
+        <ion-icon *ngIf="options?.icon" [name]="options?.icon"></ion-icon>
+        <span *ngIf="options?.title" [innerHTML]="options?.title"></span>
+    </button>
+</div>
